+++ shop/payment/urls.py
@@ -1,24 +1,15 @@
-#-*- coding: utf-8 -*-
-"""
-Loop over payment backends defined in settings.SHOP_PAYMENT_BACKENDS and add
-their URLs to the payment namespace. eg:
-http://www.example.com/shop/pay/paypal
-http://www.example.com/shop/pay/pay-on-delivery
-...
-"""
-from django.conf.urls import patterns, include
-from shop.backends_pool import backends_pool
+from django.conf.urls import patterns, url
+from shop.payment.pay_on_delivery import pay_on_delivery
+from shop.payment.credit_card import setup_view, entry_view, callback_view
 
 
-urlpatterns = patterns('')
+urlpatterns = patterns(
+	'',
+	# example URLs for simple payment
+	url(r'^payment/on-delivery/$', pay_on_delivery),
 
-# For every backend defined in the backend pool, load all the URLs it defines
-# in its get_urls() method.
-for backend in backends_pool.get_payment_backends_list():
-    regexp = '^%s/' % backend.url_namespace
-    urls = backend.get_urls()
-    pattern = patterns('',
-        (regexp, include(backend.get_urls()))
-    )
-
-    urlpatterns = pattern + urlpatterns
+	# example URLs for complicated payment backend
+	url(r'payment/credit-card/', setup_view, name="credit_card"),
+	url(r'payment/credit-card/send', entry_view, name="credit_card_entry"),
+	url(r'payment/credit-card/receive', callback_view, name="credit_card_callback"),
+)