+++ shop/util/order.py
@@ -49,3 +49,17 @@ def add_order_to_request(request, order):
         # Add the order_id to the session There has to be a session. Otherwise
         # it's fine to get an AttributeError
         request.session['order_id'] = order.pk
+
+
+def redirect_to_shipping(request, order=None):
+    if order is None:
+        order = get_order_from_request(request)
+    return HttpResponseRedirect(
+        reverse(order.shipping_backend.url_name))
+
+
+def redirect_to_payment(request, order=None):
+    if order is None:
+        order = get_order_from_request(request)
+    return HttpResponseRedirect(
+        reverse(order.payment_backend.url_name))